Updated URLconf for videos
