+++ resources/urls.py
@@ -4,6 +4,4 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-    path('videos/', views.VideoListView.as_view(), name='videos'),
-    path('v/<slug:slug>/', views.VideoDetailView.as_view(), name='video')
 ]
